requirements.txt
Flask==1.1.2
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
flask-shell-ipython=0.4.1